PR: fixed login in ProfileMenu and changed prop name in ProtectedRoute
