+++ src/pages/layout/components/header/components/profileMenu/ProfileMenu.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Tooltip, IconButton, Paper, Menu, Box } from '@mui/material';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
-import { ProfileMenuItems } from './ProfileMenuItems';
+import { ProfileMenuList } from './ProfileMenuList';
 
 export const ProfileMenu = () => {
   const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(null);
@@ -38,7 +38,7 @@ export const ProfileMenu = () => {
           open={Boolean(anchorElUser)}
           onClose={handleCloseUserMenu}
         >
-          <ProfileMenuItems closeMenu={handleCloseUserMenu} />
+          <ProfileMenuList closeMenu={handleCloseUserMenu} />
         </Menu>
       </Paper>
     </Box>
