--- src/pages/layout/components/header/components/profileMenu/constants/profileMenuList.tsx
@@ -1,24 +1,0 @@
-import PersonRoundedIcon from '@mui/icons-material/PersonRounded';
-import LogoutIcon from '@mui/icons-material/Logout';
-
-type ProfileMenuItem = {
-  name: string;
-  icon: JSX.Element;
-  navigateTo: string;
-  logout: boolean;
-};
-
-export const PROFILE_MENU_LIST: ProfileMenuItem[] = [
-  {
-    name: 'Мой профиль',
-    icon: <PersonRoundedIcon />,
-    navigateTo: '/profile',
-    logout: false,
-  },
-  {
-    name: 'Выйти',
-    icon: <LogoutIcon />,
-    navigateTo: '/login',
-    logout: true,
-  },
-];
